\name{print.cohen.kappa}
\alias{print.cohen.kappa}
\title{prints kappa reliability coefficient for nominal data}
\description{
 prints the kappa coefficient and associated probability
}
\usage{
 \method{print}{cohen.kappa}(x,...)
}
\arguments{
 \item{x}{an object of class "cohen.kappa" returned from \code{cohen.kappa}}
 \item{...}{arguments to be passed to \code{print}}
}
\value{
 nil
}
\author{Jim Lemon}
\keyword{misc}
