\name{print.kendall.w}
\alias{print.kendall.w}
\title{displays Kendall's W information}
\description{
 prints the W coefficient, associated probability and optionally a table of 
 ranks
}
\usage{
 print.kendall.w(x,...)
}
\arguments{
 \item{x}{an object of class "kendall.w" returned from \code{kendall.w}}
 \item{...}{arguments to be passed to \code{print}}
}
\value{
 nil
}
\author{Jim Lemon}
\keyword{misc}
