\name{print.kripp.alpha}
\alias{print.kripp.alpha}
\title{displays Krippendorff's alpha information}
\description{
 prints the Krippendorff's alpha coefficient
}
\usage{
 print.kripp.alpha(x,...)
}
\arguments{
 \item{x}{an object of class "kripp.alpha" returned from \code{kripp.alpha}}
 \item{...}{arguments to be passed to \code{print}}
}
\value{
 nil
}
\author{Jim Lemon}
\keyword{misc}
