\name{print.page.trend.test}
\alias{print.page.trend.test}
\title{prints kappa reliability coefficient for nominal data}
\description{
 prints the kappa coefficient and associated probability
}
\usage{
 print.page.trend.test(x,...)
}
\arguments{
 \item{x}{an object returned from \code{page.trend.test}}
 \item{...}{arguments to be passed to \code{print}}
}
\value{
 nil
}
\author{Jim Lemon}
\keyword{misc}
