\name{print.quest.reliability}
\alias{print.quest.reliability}
\title{Print questionnaire reliability measures}
\description{Prints the values returned from quest.reliability}
\usage{
 \method{print}{quest.reliability}(x,...)
}
\arguments{
 \item{x}{a list of results from quest.reliability}
 \item{...}{dummy argument to match \code{print(x,...)}}
}
\author{Jim Lemon}
\seealso{\code{\link{print}}}
\keyword{misc}
