\name{print.stuart.maxwell}
\alias{print.stuart.maxwell}
\title{displays Stuart-Maxwell marginal homogeneity test information}
\description{
 prints the information from \code{stuart.maxwell.mh}
}
\usage{
 \method{print}{stuart.maxwell}(x,...)
}
\arguments{
 \item{x}{an object of class "stuart.maxwell" returned from
  \code{stuart.maxwell}}
 \item{...}{arguments to be passed to \code{print}}
}
\value{
 nil
}
\author{Jim Lemon}
\keyword{misc}
