\name{stickleback}
\alias{stickleback}
\title{Estimates of nuptial coloration of male sticklebacks}
\description{
 Agreement of classification of 4 raters of 29 male sticklebacks into 5 
  color categories.
}
\usage{
 data(stickleback)
}
\source{
 Rowland, W. J. (1984) The relationships among nuptial coloration, aggression
 and courtship of male three-spined sticklebacks, \emph{Gasterosetus
 aculeatus}.\emph{Canadian Journal of Zoology}, \bold{62}, 999-1004.
}
\keyword{misc}
