\name{Wcrit01}
\docType{data}
\alias{Wcrit01}
\title{Critical values of Kendall's W}
\description{
 Critical values of Kendall's W (alpha=0.01)
}
\usage{data(Wcrit)}
\format{
 A matrix with 5 rows and 18 columns containing the alpha=0.01
 critical values for Kendall's W with small sample sizes.
}
\source{
 Kendall, M.G. (1970) Rank correlation methods. London: Charles Griffin.
}
\references{
 Siegel, S. & Castellan, N.J.Jr. (1988) Nonparametric
 statistics for the behavioral sciences. Boston, MA: McGraw-Hill.
}
\keyword{datasets}
