\name{Wcrit05}
\docType{data}
\alias{Wcrit05}
\title{Critical values of Kendall's W}
\description{
 Critical values of Kendall's W (alpha=0.05)
}
\usage{data(Wcrit05)}
\format{
  A data frame with 72 observations on 2 variables.
  \tabular{rlll}{
    [,1]  \tab count  \tab numeric  \tab Insect count\cr
    [,2]  \tab spray  \tab factor   \tab The type of spray
  }
}
\source{
 Kendall, M.G. (1970) Rank correlation methods. London: Charles Griffin.
}
\references{
 Siegel, S. & Castellan, N.J.Jr. (1988) Nonparametric
 statistics for the behavioral sciences. Boston, MA: McGraw-Hill.
}
\keyword{datasets}
